moved QRCode Image genarator to UIImage extension for reuse
